aia/bicep/requirements.txt
akshare==1.9.67
attrs==23.1.0
backtrader==1.9.78.123
beautifulsoup4==4.12.2
cattrs==22.2.0
certifi==2022.12.7
cfscrape==2.1.1
charset-normalizer==3.1.0
chevron==0.14.0
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fonttools==4.39.3
html5lib==1.1
idna==3.4
jsonpath==0.82
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
platformdirs==3.5.0
py-mini-racer==0.6.0
pyparsing==3.0.9
pypinyin==0.48.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML @ file:///Users/runner/miniforge3/conda-bld/pyyaml_1692737308199/work
requests==2.29.0
requests-cache==1.0.1
six==1.16.0
soupsieve==2.4.1
tabulate==0.9.0
tqdm==4.65.0
tzdata==2023.3
url-normalize==1.4.3
urllib3==1.26.15
webencodings==0.5.1
websocket-client==1.5.1
xlrd==2.0.1
